Set background of cells with no value
I had to use my own fork of cal-heatmap to fix this. The `empty` flag
does not work for cells containing no data (see cal-heatmap
[documentation](http://kamisama.github.io/cal-heatmap/#legendColors))

I believe this is a bug in cal-heatmap and fixed it on a local fork and
sent a [pull request](https://github.com/kamisama/cal-heatmap/pull/74)
some time ago. I now updated this pull request so it can again be
automatically merged.