Define number of domains based on svg width
The number of domains rendered in the year chart is defined based on
the width of the svg element. This was mentioned in #43. The number of
domains is calculated using a heuristic. It might not be 100% correct
but since the width of the svg element can only be one of a limited set
of allowed widths, this solution will always render the optimal number
of domains.